cleanup - gonna scale this back and get rid of the localization framework
It's too bothersome and *way* beyond the scope of a microframework.

This framework is only about 2 things: Creating better Flash sites with
deeplinking and a proper/easy to use fallback system to generate HTML
alternatives - by hand. No auto-generating that will only be useful to
some, but difficult to modify for others.

Make this bareboned, and then rely on subsequent implementers to choose
their own strategy for dynamic content, localization, etc…